requirements.txt
requests
schedule
python-dotenv
yagmail
MIMEText
MIMEMultipart
smtplib
